fix: audit - version long, cifrado default, no plaintext fallback, cleanup
- ClaveGrupo/ClaveGrupoRequest/ClaveGrupoDAO: int version -> long (Date.now() overflows int)
- Mensaje.cifrado: add columnDefinition default false (NOT NULL ADD COLUMN on existing table)
- app.js: abort with error instead of silent plaintext fallback in all 3 send paths
- crypto.js decryptFile: remove dead export+re-import (importGroupKey already usable)
- ClaveGrupoDAO: remove unused import java.util.List
- ClaveGrupoController: split catch so token failure -> 401, persist failure -> 500

+++ src/main/java/com/chat/dao/ClaveGrupoDAO.java
@@ -6,8 +6,6 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.NoResultException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
-import java.util.List;
-
 @ApplicationScoped
 public class ClaveGrupoDAO {
 
@@ -16,7 +14,7 @@ public class ClaveGrupoDAO {
 
     @Transactional
     public void guardarOActualizar(int chatId, int miembroId, int distribuidorId,
-                                   String claveEnvuelta, int version) {
+                                   String claveEnvuelta, long version) {
         em.createQuery(
             "DELETE FROM ClaveGrupo c WHERE c.chatId = :chatId AND c.miembroId = :miembroId")
           .setParameter("chatId", chatId)
